test: separate logic for utils
